Fix reading data from different classes
When an inherited class of lazyrecord is defined, the inherited class method is triggered, opening the store with all the objects. However, by that time not all the class definitions were loaded, raising an exception.

The inherited method was replaced with .register(*classes). We must ensure to call this method before using the store.